Database access improvements on the backend
Edited 'DatabaseFunctions' so it now features a fully capable
'read(Table table)' method. The data returned is a dictionary:
'Dictionary<string, List<Object>>' where for key "Names" this
list is a list of column names, and for key "Values" this List
will represent a list of rows which are also a List by themselves.

For this feature implementation Model improvements were required.
Now all Table Models return their specific name (_id), and this
information is used by the reader to identify which table to read
from. Constructors were also added too.

'databaseTest()' method was adapted for this new change.