Added PostAd HTTPS method
PostAd method enables making new posts to the website. PostAd
expects same the JSON format the server passes to the client.
Returns 200 if successful, 400 if not.

WARNING: post failure cleanup still not implemented, this can lead
to "uncompleted" (dirty) new posts in the database. Keep possible
dirty posts in mind.